\name{SimData}
\alias{SimData}
\docType{data}
\title{Simulated data with two classes}
\description{
Simulated longitudinal data with 10 measured time points for 20 subjects.  The subjects (rows) belong to two classes: the first 10 have a trajectory with one changepoint, the next 10 have trajevtory with two changepoints.       
}
\usage{data(SimData)}
\format{
\itemize{
\item X: 20 X 10 matrix of time points for each subject  
\item Y: 20 X 10 matrix of outcomes measurements for each subject
}}

\keyword{datasets}
