\name{SimData4classes}
\alias{SimData4classes}
\docType{data}
\title{Simulated data with four classes}
\description{
Simulated longitudinal data with 10 measured time points for 40 subjects.  The subjects (rows) belong to four classes: the first 10 have a trajectory with zero changepoints, the next 10 have trajevtory with one changepoint, the next ten have a trajectory with two changepoints, and the next 10 have a trajectory with three changepoints.       
}
\usage{data(SimData4classes)}
\format{
\itemize{
\item X: 40 X 10 matrix of time points for each subject  
\item Y: 40 X 10 matrix of outcomes measurements for each subject
}}

\keyword{datasets}
